% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.move.data.R
\name{get.move.data}
\alias{get.move.data}
\title{Import Movebank study location and reference data into R}
\usage{
get.move.data(location, reference, remove_movebank_outliers, omit_derived_data)
}
\arguments{
\item{project}{Character string. Name of the project folder.}

\item{path}{Character string. Directory path where the cloned project will be located. Default is the working directory, getwd().}

\item{main}{Logical. Whether to return results from the main project folder (all subfolders except "incoming"). Default is TRUE.}

\item{incoming}{Logical. Whether to return results from the "incoming" project subfolder. Default is TRUE.}
}
\value{
Returns a download of the project folder and all its contents.
}
\description{
Imports location and reference data from one or more Movebank studies managed by a specified Movebank account.
}
\examples{
# e.g.import <- get.move.data(location = TRUE, reference = TRUE, remove_movebank_outliers <- FALSE, omit_derived_data <- FALSE)

}
\seealso{
\verb{ }
}
\keyword{import,}
\keyword{movebank,}
\keyword{tracking}
