% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importer.R
\name{importer}
\alias{importer}
\title{Import one or more files into R}
\usage{
importer(
  path,
  multiple = FALSE,
  graphics = TRUE,
  pattern = NULL,
  column.format = list(c(column = NULL, format = NULL))
)
}
\arguments{
\item{path}{Character string. Directory path here files are located. Default is the working directory, getwd().}

\item{multiple}{Logical. Whether to allow multiple file selections. This will always return an R list object. Default is FALSE.}

\item{graphics}{Logical. Whether to use a graphical widget for file selection. Default is TRUE.}

\item{pattern}{Character vector. File name pattern(s). Must be a regular expression; print ?base::regex for help. Default is NULL, which returns results for all files.}

\item{column.format}{Optional list. Specifies formatting for one or more data frame columns, particularly those of class "Date."}
}
\value{
Returns an R list object (multiple = TRUE or the imported file is json format) or data frame (when multiple = FALSE and the imported file is not json format).
}
\description{
Supports csv, Excel, json, gpx, kml, and/or kmz files.
}
\examples{
# e.g.import <- importer(path = "downloads/, multiple = FALSE, graphics = TRUE)

}
\seealso{
\verb{ }
}
\keyword{csv,}
\keyword{import,}
\keyword{json}
\keyword{xls,}
\keyword{xlsx,}
