% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector.R
\name{selector}
\alias{selector}
\title{Make a selection from a set of choices}
\usage{
selector(choices, multiple = FALSE, graphics = TRUE, title = NULL)
}
\arguments{
\item{choices}{Character vector.}

\item{multiple}{Logical. Whether to allow multiple selections. Default is FALSE.}

\item{graphics}{Logical. Whether to use a graphical widget for file selection. Default is TRUE.}

\item{title}{Character string. Default is NULL.}
}
\value{
Returns character string (multiple = FALSE) or vector (multiple = TRUE) of choice selection.
}
\description{
Make a selection from a set of choices
}
\examples{
# 

}
\seealso{
\verb{ }
}
\keyword{filter}
