% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importer.R
\name{importer}
\alias{importer}
\title{Import one or more files into R}
\usage{
importer(path, multiple, graphics, column.format = list(c(column, format)))
}
\arguments{
\item{path}{Character string. Directory path here files are located. Default is the working directory, getwd().}

\item{multiple}{Logical. Whether to allow multiple file selections. This will always return an R list object. Default is FALSE.}

\item{graphics}{Logical. Whether to use a graphical widget for file selection. Default is TRUE.}

\item{formats}{Optional list. Specifies formatting for one or more data frame columns, particularly those of class "Date."}
}
\value{
Returns an R list object (multiple = TRUE or the imported file is json format) or data frame (when multiple = FALSE and the imported file is not json format).
}
\description{
Supports csv, Excel, and/or json files.
}
\examples{
# e.g.import <- importer(path = "downloads/, multiple = FALSE, graphics = TRUE)

}
\seealso{
\verb{ }
}
\keyword{csv,}
\keyword{import,}
\keyword{json}
\keyword{xls,}
\keyword{xlsx,}
